+++ pkg/virt-operator/resource/generate/components/validatingadmissionpolicies/validatingadmissionpolicy_handler.go
@@ -17,7 +17,7 @@
  *
  */
 
-package components
+package validatingadmissionpolicies
 
 import (
 	"fmt"
